Happify mypy

+++ test/test_ir.py
@@ -42,7 +42,7 @@ class TestAxis:
         ],
     )
     def test_axis_setter_no_error(self, axis: Axis, new_axis: ArrayLike, expected_axis: ArrayLike) -> None:
-        axis.value = new_axis
+        axis.value = new_axis  # type: ignore[assignment]
         np.testing.assert_array_equal(axis, expected_axis)
 
     @pytest.mark.parametrize(
@@ -278,12 +278,12 @@ class TestMatrixGate:
         assert gate.get_qubit_operands() == [Qubit(42), Qubit(100)]
 
     def test_is_identity(self, gate: MatrixGate) -> None:
-        assert MatrixGate(np.eye(4), operands=[Qubit(42), Qubit(100)]).is_identity()
+        assert MatrixGate(np.eye(4, dtype=np.complex128), operands=[Qubit(42), Qubit(100)]).is_identity()
         assert not gate.is_identity()
 
     def test_matrix_gate_same_control_and_target_qubit(self) -> None:
         with pytest.raises(ValueError) as e_info:
-            MatrixGate(np.eye(4), [Qubit(0), Qubit(0)])
+            MatrixGate(np.eye(4, dtype=np.complex128), [Qubit(0), Qubit(0)])
 
         assert "control and target qubit cannot be the same" in str(e_info.value)
